Avoid calling `getRegexForFormatString` for `display` parameters
Fixes #94

There’s a bug that makes a `display` string, which is meant solely for
appearing in Hubot’s help, to go through the same transformations as
normal commands, which means it will be converted to a regex. It’s not
just unnecessary, but actually causes an error in some cases.