Add basic 2fa support [WIP]
Add support for two-factor auth: the `extra.security.twofactor`
parameter is checked in the alias and, if present, the user is asked to
run the auth alias with a random guid.
This support relies on `2fa {{ uuid }}` alias being present and also on
the successful 2fa attempts emitting `2fa` events.