Remove undocumented ST2_AUTH_TOKEN from st2chatops
`ST2_AUTH_TOKEN` is not documented in `st2chatops.env` and because the token either has a long expiry time or the expiry time is so short the ChatOps service is restarted all the time to pick up the config changes. It is no point to keep it, therefore remove this auth method from hubot-stackstorm entirely.